domain type is now aware of admin selected domains
