+++ packages/framework/src/Form/DomainType.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Shopsys\FrameworkBundle\Form;
 
+use Shopsys\FrameworkBundle\Component\Domain\AdminDomainTabsFacade;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\IntegerType;
@@ -15,24 +16,27 @@ class DomainType extends AbstractType
 {
     /**
      * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
+     * @param \Shopsys\FrameworkBundle\Component\Domain\AdminDomainTabsFacade $adminDomainTabsFacade
      */
-    public function __construct(private readonly Domain $domain)
-    {
+    public function __construct(
+        private readonly Domain $domain,
+        private readonly AdminDomainTabsFacade $adminDomainTabsFacade,
+    ) {
     }
 
     /**
      * {@inheritdoc}
      */
-    public function buildView(FormView $view, FormInterface $form, array $options)
+    public function buildView(FormView $view, FormInterface $form, array $options): void
     {
-        $view->vars['domainConfigs'] = $this->domain->getAll();
+        $view->vars['domainConfigs'] = $this->getSortedDomainConfigsByAdminDomainTabs();
         $view->vars['displayUrl'] = $options['displayUrl'];
     }
 
     /**
      * @param \Symfony\Component\OptionsResolver\OptionsResolver $resolver
      */
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'displayUrl' => false,
@@ -40,9 +44,28 @@ class DomainType extends AbstractType
     }
 
     /**
+     * @return \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig[]
+     */
+    private function getSortedDomainConfigsByAdminDomainTabs(): array
+    {
+        $selectedDomainId = $this->adminDomainTabsFacade->getSelectedDomainId();
+
+        $list = [];
+        $list[] = $this->adminDomainTabsFacade->getSelectedDomainConfig();
+
+        foreach ($this->domain->getAdminEnabledDomains() as $domainConfig) {
+            if ($domainConfig->getId() !== $selectedDomainId) {
+                $list[] = $domainConfig;
+            }
+        }
+
+        return $list;
+    }
+
+    /**
      * {@inheritdoc}
      */
-    public function getParent(): ?string
+    public function getParent(): string
     {
         return IntegerType::class;
     }
